pyproject.toml
[project]
name = "chat-agent"
version = "0.1.0"
description = "Production-ready AI agent chat API"
requires-python = "==3.11.*"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0,<0.31.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "anthropic>=0.26.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.9",
    "structlog>=24.1.0",
    "greenlet>=3.0.0,<=3.2.5",
    "mangum>=0.21.0",
    "requests<=2.26,>=2.20.1",
    "six<1.15.0,>=1.11.0",
    "jsonschema<4,>=2.5.1",
    "PyYAML>=6.0",
    "boto3>=1.42.96",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
